vgabios: Minor - pass display address to stdvga_set_cursor_pos().

+++ vgasrc/stdvga.h
@@ -131,7 +131,7 @@ void stdvga_load_font(u16 seg, void *src_far, u16 count
 u16 stdvga_get_crtc(void);
 int stdvga_bpp_factor(struct vgamode_s *vmode_g);
 void stdvga_set_cursor_shape(u8 start, u8 end);
-void stdvga_set_cursor_pos(u16 address);
+void stdvga_set_cursor_pos(int address);
 void stdvga_set_scan_lines(u8 lines);
 u16 stdvga_get_vde(void);
 int stdvga_get_window(struct vgamode_s *vmode_g, int window);
